Ultima actualizacion del parser

+++ lectura.cpp
@@ -1,6 +1,6 @@
 #include "lectura.h"
 
-Lectura::Lectura(string titulo, double minutos, int anio, Escritor escritor) {
+Lectura::Lectura(string titulo, float minutos, int anio, Escritor escritor) {
     this->titulo = titulo;
     this->minutos = minutos;
     this->anio = anio;
