+++ lectura.h
@@ -7,14 +7,14 @@ class Lectura {
 
 protected:
     string titulo;
-    double minutos;
+    float minutos;
     int anio;
     Escritor escritor;
 
 public:
     //PRE:
     //POST: Crea una nueva lectura
-    Lectura(string titulo, double minutos, int anio, Escritor escritor);
+    Lectura(string titulo, float minutos, int anio, Escritor escritor);
 
     //PRE:
     //POST: Devuelve el anio
