+++ libro.h
@@ -7,7 +7,7 @@ class Libro : public Lectura{
     public:
     //Pre:
     //Post: Crea un libro nuevo.
-    Libro(string titulo, double minutos, int anio, Escritor escritor, string titulo_libro);
+    Libro(string titulo, float minutos, int anio, Escritor escritor, string titulo_libro);
 
     //Pre:
     //Post: Muestra un la informacion de un libro completo por pantalla.
