+++ novela_historica.cpp
@@ -1,6 +1,6 @@
 #include "novela_historica.h"
 
-Novela_historica::Novela_historica(string titulo, double minutos, int anio, Escritor escritor, int genero, string tema) 
+Novela_historica::Novela_historica(string titulo, float minutos, int anio, Escritor escritor, int genero, string tema) 
                             : Novela(titulo, minutos, anio, escritor, HISTORICA){
