+++ novela_historica.h
@@ -13,7 +13,7 @@ public:
     
     //Pre:
     //Post: Crea una novela_historica nueva.
-    Novela_historica(string titulo, double minutos, int anio, Escritor escritor, int genero, string tema);
+    Novela_historica(string titulo, float minutos, int anio, Escritor escritor, int genero, string tema);
 
     //Pre:
     //Post: Muestra la informacion de una novela_historica.
